move integration tests to proper location
